Moved more code into AbstractEnemy.cs, so that it can be shared, and doesn't have to be reimplmlemented.
